+++ src/main/java/net/jagunma/backbone/auth/authmanager/infrastructure/controller/web/oa11010/vo/Oa11010BizTranRoleVo.java
@@ -11,7 +11,7 @@ public class Oa11010BizTranRoleVo {
 	/**
 	 * 取引ロール選択
 	 */
-	private Integer bizTranRoleSelected;
+	private Short bizTranRoleSelected;
 	/**
 	 * 取引ロールID
 	 */
@@ -53,8 +53,8 @@ public class Oa11010BizTranRoleVo {
 	 */
 	private LocalDate expirationEndDateTo;
 
-	public Integer getBizTranRoleSelected() { return bizTranRoleSelected; }
-	public void setBizTranRoleSelected(Integer bizTranRoleSelected) { this.bizTranRoleSelected = bizTranRoleSelected; }
+	public Short getBizTranRoleSelected() { return bizTranRoleSelected; }
+	public void setBizTranRoleSelected(Short bizTranRoleSelected) { this.bizTranRoleSelected = bizTranRoleSelected; }
 	public long getBizTranRoleId() { return bizTranRoleId; }
 	public void setBizTranRoleId(long bizTranRoleId) { this.bizTranRoleId = bizTranRoleId; }
 	public String getBizTranRoleCode() { return bizTranRoleCode; }
