+++ src/main/java/net/jagunma/backbone/auth/authmanager/infrastructure/controller/web/oa11010/vo/Oa11010SubSystemRoleVo.java
@@ -12,7 +12,7 @@ public class Oa11010SubSystemRoleVo implements Serializable {
 	/**
 	 * サブシステムロール選択
 	 */
-	private Integer subSystemRoleSelected;
+	private Short subSystemRoleSelected;
 	/**
 	 * サブシステムロールID
 	 */
@@ -50,8 +50,8 @@ public class Oa11010SubSystemRoleVo implements Serializable {
 	 */
 	private LocalDate expirationEndDateTo;
 
-	public Integer getSubSystemRoleSelected() { return subSystemRoleSelected; }
-	public void setSubSystemRoleSelected(Integer subSystemRoleSelected) { this.subSystemRoleSelected = subSystemRoleSelected; }
+	public Short getSubSystemRoleSelected() { return subSystemRoleSelected; }
+	public void setSubSystemRoleSelected(Short subSystemRoleSelected) { this.subSystemRoleSelected = subSystemRoleSelected; }
 	public long getSubSystemRoleId() { return subSystemRoleId; }
 	public void setSubSystemRoleId(long subSystemRoleId) { this.subSystemRoleId = subSystemRoleId; }
 	public String getSubSystemRoleCode() { return subSystemRoleCode; }
